fix(rpcserver): use default NIO instead EPOLL for ARM compatibility
nio's EPoll support only works for AMD64 platforms. On any other platform,
it just fails. We settle back to normal nio so it also works with other
platforms out of the box. As their are only a few local connections,
using EPoll should not have provided much of a performance gain.
So, it should be fine to use default nio.